requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.9.3
aiosignal==1.3.2
airportsdata==20250224
altair==5.3.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.22.1
bitsandbytes==0.45.4
blake3==1.0.4
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
ccxt==4.4.61
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
compressed-tensors==0.9.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
depyf==0.18.0
dill==0.3.9
discord.py==2.5.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
duckdb==0.10.0
einops==0.8.1
email_validator==2.2.0
exchange_calendars==4.9
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==24.3.25
fonttools==4.56.0
frozendict==2.4.4
frozenlist==1.5.0
fsspec==2024.6.0
gast==0.5.4
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
# Editable install with no version control (grokapi==0.1.0)
# -e c:\users\shoaib\cascadeprojects\grok3-api # Commented out - might not be standard package format
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==7.1.0
iniconfig==2.0.0
injector==0.20.1
intel-openmp==2021.4.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.6.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keract==4.5.1
keras==3.3.3
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
lark==1.2.2
libclang==18.1.1
lm-format-enforcer==0.10.11
loguru==0.6.0
lxml==4.9.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mistral_common==1.5.4
mkl==2021.4.0
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
msgspec==0.19.0
multidict==6.1.0
multitasking==0.0.11
namex==0.0.8
nbeats-pytorch==1.8.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.65.4
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.1
pandas_market_calendars==4.6.1
pandas_ta==0.3.14b0
partial-json-parser==0.2.1.1.post5
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
playwright==1.50.0
plotly==5.6.0
pluggy==1.5.0
polygon-api-client==1.12.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycares==4.5.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyee==12.1.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyluach==2.2.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-telegram-bot==21.4
pytweening==1.2.0
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.3.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-futures==1.0.2
requests-oauthlib==2.0.0
retrying==1.3.3
rich==13.7.1
rich-toolkit==0.13.2
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
schedule==1.2.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.25.0
semver==3.0.2
sentencepiece==0.2.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.46.1
StrEnum==0.4.15
sympy==1.12.1
ta==0.11.0
tastytrade==9.11
# tastytrade-sdk==1.0.0 # Removed due to newer tastytrade package
tavily-python==0.5.1
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.50.3
trio==0.27.0
trio-websocket==0.11.1
tweepy==4.15.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
undetected-chromedriver==3.5.5
unusualwhales-python-client==5.0.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
validators==0.28.3
vllm==0.8.1
watchdog==4.0.1
watchfiles==1.0.4
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.16.0
wsproto==1.2.0
yahooquery==2.3.7
yarl==1.18.3
yfinance==0.2.37
zipp==3.19.2
